+++ analytics.js
@@ -1046,13 +1046,14 @@ analytics.addProvider('Google Analytics', {
 
         // Try to check for a `category` and `label`. A `category` is required,
         // so if it's not there we use `'All'` as a default. We can safely push
-        // undefined if the special properties don't exist.
+        // undefined if the special properties don't exist. Try using revenue
+        // first, but fall back to a generic `value` as well.
         window._gaq.push([
             '_trackEvent',
             properties.category || 'All',
             event,
             properties.label,
-            value,
+            Math.round(properties.revenue) || value,
             properties.noninteraction
         ]);
     },
@@ -1489,6 +1490,12 @@ analytics.addProvider('KISSmetrics', {
     // -----
 
     track : function (event, properties) {
+        // KISSmetrics handles revenue with the `'Billing Amount'` property by
+        // default, although it's changeable in the interface.
+        analytics.utils.alias(properties, {
+            'revenue' : 'Billing Amount'
+        });
+
         window._kmq.push(['record', event, properties]);
     }